+++ expr_sema.go
@@ -322,7 +322,11 @@ var BuiltinGlobalVariableTypes = map[string]ExprType{
 				"ports":   NewMapObjectType(StringType{}),
 			}),
 		),
-		"status": StringType{},
+		"status":              StringType{},
+		"workflow_file_path":  StringType{}, // https://docs.github.com/en/actions/reference/workflows-and-actions/contexts#job-context
+		"workflow_ref":        StringType{},
+		"workflow_repository": StringType{},
+		"workflow_sha":        StringType{},
 	}),
 	// https://docs.github.com/en/actions/learn-github-actions/contexts#steps-context
 	"steps": NewEmptyStrictObjectType(), // This value will be updated contextually