+++ linter.go
@@ -561,6 +561,7 @@ func (l *Linter) check(
 			NewRuleRunnerLabel(),
 			NewRuleEvents(),
 			NewRuleJobNeeds(),
+			NewRuleParallelSteps(),
 			NewRuleAction(localActions),
 			NewRuleEnvVar(),
 			NewRuleID(),